change: Use TypeScript for the fastxml http transform
Besides adding type information to the fastxml module the tests got
refactored and the parseEnvelopeMeta function got moved from the parser
module to it's only user the fastxml module. The parseEnvelopeMeta
function is also private now.
