+++ src/gmp/__tests__/parser.test.js
@@ -10,7 +10,6 @@ import {
   parseBoolean,
   parseDate,
   parseDuration,
-  parseEnvelopeMeta,
   parseFloat,
   parseInt,
   parseProgressElement,
@@ -256,50 +255,6 @@ describe('parseQod tests', () => {
     ).toEqual({
       value: 55,
       type: 'remote_vul',
-    });
-  });
-});
-
-describe('parseEnvelopeMeta tests', () => {
-  test('should parse envelope information', () => {
-    expect(
-      parseEnvelopeMeta({
-        version: '1.0',
-        backend_operation: '0.01',
-        vendor_version: '1.1',
-        i18n: 'en',
-        time: 'Fri Sep 14 11:26:40 2018 CEST',
-        timezone: 'Europe/Berlin',
-      }),
-    ).toEqual({
-      version: '1.0',
-      backendOperation: '0.01',
-      vendorVersion: '1.1',
-      i18n: 'en',
-      time: 'Fri Sep 14 11:26:40 2018 CEST',
-      timezone: 'Europe/Berlin',
-    });
-  });
-
-  test('should drop unknown envelope information', () => {
-    expect(
-      parseEnvelopeMeta({
-        version: '1.0',
-        backend_operation: '0.01',
-        vendor_version: '1.1',
-        i18n: 'en',
-        time: 'Fri Sep 14 11:26:40 2018 CEST',
-        timezone: 'Europe/Berlin',
-        foo: 'bar',
-        lorem: 'ipsum',
-      }),
-    ).toEqual({
-      version: '1.0',
-      backendOperation: '0.01',
-      vendorVersion: '1.1',
-      i18n: 'en',
-      time: 'Fri Sep 14 11:26:40 2018 CEST',
-      timezone: 'Europe/Berlin',
     });
   });
 });
